tls: splitting config into downstream and upstream
