+++ source/extensions/transport_sockets/tls/config.h
@@ -18,30 +18,6 @@ public:
   std::string name() const override { return "envoy.transport_sockets.tls"; }
 };
 
-class UpstreamSslSocketFactory : public Server::Configuration::UpstreamTransportSocketConfigFactory,
-                                 public SslSocketConfigFactory {
-public:
-  Network::UpstreamTransportSocketFactoryPtr createTransportSocketFactory(
-      const Protobuf::Message& config,
-      Server::Configuration::TransportSocketFactoryContext& context) override;
-  ProtobufTypes::MessagePtr createEmptyConfigProto() override;
-};
-
-DECLARE_FACTORY(UpstreamSslSocketFactory);
-
-class DownstreamSslSocketFactory
-    : public Server::Configuration::DownstreamTransportSocketConfigFactory,
-      public SslSocketConfigFactory {
-public:
-  Network::DownstreamTransportSocketFactoryPtr
-  createTransportSocketFactory(const Protobuf::Message& config,
-                               Server::Configuration::TransportSocketFactoryContext& context,
-                               const std::vector<std::string>& server_names) override;
-  ProtobufTypes::MessagePtr createEmptyConfigProto() override;
-};
-
-DECLARE_FACTORY(DownstreamSslSocketFactory);
-
 } // namespace Tls
 } // namespace TransportSockets
 } // namespace Extensions
