+++ source/extensions/transport_sockets/tls/upstream_config.cc
@@ -1,4 +1,4 @@
-#include "source/extensions/transport_sockets/tls/config.h"
+#include "source/extensions/transport_sockets/tls/upstream_config.h"
 
 #include "envoy/extensions/transport_sockets/tls/v3/cert.pb.h"
 #include "envoy/extensions/transport_sockets/tls/v3/tls.pb.validate.h"
@@ -30,26 +30,6 @@ ProtobufTypes::MessagePtr UpstreamSslSocketFactory::createEmptyConfigProto() {
 
 LEGACY_REGISTER_FACTORY(UpstreamSslSocketFactory,
                         Server::Configuration::UpstreamTransportSocketConfigFactory, "tls");
-
-Network::DownstreamTransportSocketFactoryPtr
-DownstreamSslSocketFactory::createTransportSocketFactory(
-    const Protobuf::Message& message, Server::Configuration::TransportSocketFactoryContext& context,
-    const std::vector<std::string>& server_names) {
-  auto server_config = std::make_unique<ServerContextConfigImpl>(
-      MessageUtil::downcastAndValidate<
-          const envoy::extensions::transport_sockets::tls::v3::DownstreamTlsContext&>(
-          message, context.messageValidationVisitor()),
-      context);
-  return std::make_unique<ServerSslSocketFactory>(
-      std::move(server_config), context.sslContextManager(), context.statsScope(), server_names);
-}
-
-ProtobufTypes::MessagePtr DownstreamSslSocketFactory::createEmptyConfigProto() {
-  return std::make_unique<envoy::extensions::transport_sockets::tls::v3::DownstreamTlsContext>();
-}
-
-LEGACY_REGISTER_FACTORY(DownstreamSslSocketFactory,
-                        Server::Configuration::DownstreamTransportSocketConfigFactory, "tls");
 
 } // namespace Tls
 } // namespace TransportSockets
